feat(v2 spec): version 2 of the Protocol
Protocol is also refered to as "maidenlane", which is compromised of AS2 non-repudiation messaging
and the Etheruem Wire p2p. This specification is implemented trhough the Hyperledger Besu client and
an Active MQ system for internal interconnections. This diagram represents the interface components
between various subsystems the protocol interacts with.